DVcon-main/requirements.txt
torch
torchvision
ultralytics
opencv-python
matplotlib
numpy
tqdm
Pillow
transformers
